+++ skymap.cpp
@@ -341,7 +341,7 @@ static void smRenderUSNO2Stars(mapView_t *mapView, CSkPainter *pPainter, int reg
     if (zone == NULL)
       return;
 
-    long *pData = zone->pData;
+    qint32 *pData = zone->pData;
 
     for (int i = 0; i < zone->starCount; i++)
     {
@@ -1356,7 +1356,6 @@ static void renderSatellites(mapView_t *mapView, CSkPainter *pPainter)
           QRect rc2 = QRect(-5, -8, 10, -20);
           QRect rc3 = QRect(-5, 8, 10, 20);
 
-
           pPainter->save();
           pPainter->translate(pt.sx, pt.sy);
           pPainter->scale(0.5 * g_skSet.map.satellite.size, 0.5 * g_skSet.map.satellite.size);
@@ -1369,7 +1368,6 @@ static void renderSatellites(mapView_t *mapView, CSkPainter *pPainter)
           pPainter->drawLine(10, 0, 25, 0);
 
           pPainter->restore();
-
 
           //pPainter->setBrush(QColor(255, 0, 0));
           //pPainter->drawEllipse(QPoint(pt.sx, pt.sy), 5, 5);
